+++ app.py
@@ -1,84 +1,21 @@
 from flask import Flask, render_template
-import random 
 from route import bp
+
+def create_app():
+    app = Flask(__name__)
+
+    app.register_blueprint(bp)
+    return app
 
 app = Flask(__name__)
 
 IMG_DIR = './static'
 
-@bp.route('/combat')
-def dice_roll() -> str:
-    numbers = random.randint(1, 20)
-
-    if numbers == 1:
-        return 'critical miss!'
-    if numbers == 20:
-        return 'critical hit!'
-    else:
-        return 'you roll for ' + str(numbers) + ' damage!'
-    
-@bp.route('/advantage')
-def adv_endpoint() -> str:
-    dice_one = random.randint(1, 20)
-    dice_two = random.randint(1, 20)
-    
-    if dice_one > dice_two:
-        if dice_one == 1:
-            return 'critical miss!'
-        if dice_one == 20:
-            return 'critical hit!'
-        else:
-            return 'you roll for ' + str(dice_one) + ' damage!'
-        
-    if dice_two > dice_one:
-        if dice_two == 1:
-            return 'critical miss!'
-        if dice_two == 20:
-            return 'critical hit!'
-        else:
-            return 'you roll for ' + str(dice_two) + ' damage!'
-
-    else:
-        if dice_one == 1:
-            return 'critical miss!'
-        if dice_one == 20:
-            return 'critical hit!'
-        else:
-            return 'you roll for ' + str(dice_one) + ' damage!'
-
-@bp.route('/disadvantage')
-def dis_endpoint() -> str:
-    dice_one = random.randint(1, 20)
-    dice_two = random.randint(1, 20)
-    
-    if dice_one < dice_two:
-        if dice_one == 1:
-            return 'critical miss!'
-        if dice_one == 20:
-            return 'critical hit!'
-        else:
-            return 'you roll for ' + str(dice_one) + ' damage!'
-        
-    if dice_two < dice_one:
-        if dice_two == 1:
-            return 'critical miss!'
-        if dice_two == 20:
-            return 'critical hit!'
-        else:
-            return 'you roll for ' + str(dice_two) + ' damage!'
-
-    else:
-        if dice_one == 1:
-            return 'critical miss!'
-        if dice_one == 20:
-            return 'critical hit!'
-        else:
-            return 'you roll for ' + str(dice_one) + ' damage!'
-
-@bp.route('/image')
+@app.route('/image')
 def serve_image():
     "a simple HTTP image"
     return render_template('image.html')
 
 if __name__ == '__main__':
+    app = create_app()
     app.run(host='localhost', port=8081)
