+++ test_app.py
@@ -1,8 +1,8 @@
 import pytest
 
-from route import bp 
-from flask import Flask
 from app import create_app
+from route import bp
+from flask import Flask
 
 
 @pytest.fixture
